+++ age3-core/src/main/java/pl/edu/agh/age/compute/api/topology/package-info.java
@@ -18,11 +18,8 @@
  */
 
 /**
- * Base imports and definitions for JS DSL
- **/
-
-var List = Java.type('javaslang.collection.List');
-var HashMap = Java.type('javaslang.collection.HashMap');
-var TimedStopCondition = Java.type('pl.edu.agh.age.compute.stream.TimedStopCondition');
-var NoOpLoggingService = Java.type('pl.edu.agh.age.compute.stream.logging.NoOpLoggingService');
-var Topology = Java.type('pl.edu.agh.age.compute.api.topology.Topology');
+ * Topology API for "compute".
+ *
+ * @see pl.edu.agh.age.compute.api.TopologyProvider
+ */
+package pl.edu.agh.age.compute.api.topology;